Implement customCheckbocks and refactor CheckboxesCustomTests to CustomCheckboxTests with new implementation
